chore: app refactor
Break out models, services, routes, forms
