+++ .gitignore
@@ -13,8 +13,11 @@ dist/
 build/
 
 # app
-instance
-static
+*.db
+/archive/sendes/static/*
+/sendes/app/static/*
+
+!/sendes/app/static/favicon/
 
 # OS
 .DS_Store
