+++ .vscode/launch.json
@@ -10,12 +10,13 @@
             "request": "launch",
             "module": "flask",
             "env": {
-                "FLASK_APP": "operate/app",
-                "FLASK_DEBUG": "1"
+                "FLASK_APP": "sendes/app",
+                "FLASK_ENV": "development",
+                "FLASK_DEBUG": "1",
+                "SECRET_KEY": "ok"
             },
             "args": [
-                "run",
-                "--no-reload"
+                "run"
             ],
             "jinja": true,
             "justMyCode": true
