+++ sendes/README.md
@@ -9,4 +9,10 @@ Returns list of all constants with a button for each row allowing activation, an
 ### Get
 Returns constant creation form with buttons for creating the list of constants.
 
-##
+## User Stories
+### Test Conductor
+As a Test Conductor or Operator, I want to configure and run tests of the LRE Orifice Calibration device so
+the CD can be found quickly and accurately.
+
+### Customer
+As a Customer, I want to find an accurate CD for a given orifice.
