+++ sendes/app/templates/base.html
@@ -4,6 +4,7 @@
   {% block head %}
   <!-- Bootstrap CSS -->
   <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" rel="stylesheet">
+  <link rel="shortcut icon" href="{{ url_for('main.favicon') }}">
   
   <title>{% block title %}{% endblock %} - LRE SenDes</title>
   {% endblock %}
