+++ sendes/requirements.txt
@@ -3,7 +3,9 @@ flask_sqlalchemy==3.1.1
 flask_wtf==1.2.1
 labjack_ljm==1.21.0
 matplotlib==3.8.0
-numpy==1.26.1
+numpy==1.26.2
+Pint==0.22
+rich==13.7.0
 SQLAlchemy==2.0.21
-WTForms==3.1.0
 uncertainties==3.1.7
+WTForms==3.1.1
